requirements.txt
streamlit==1.48.0
pandas==2.2.1
gspread==6.0.0
gspread-dataframe==3.3.1
google-auth==2.35.0
google-auth-oauthlib==1.2.3
oauth2client==4.1.3
openpyxl==3.1.2
requests==2.31.0
python-dotenv==1.0.1
plotly==6.4.0
protobuf==5.28.3
